+++ packages/apps/public/locales/en/app-bounties.json
@@ -1,4 +1,5 @@
 {
+  "Accept": "Accept",
   "Account does not have enough funds.": "Account does not have enough funds.",
   "Add Bounty": "Add Bounty",
   "Allocation value is smaller than the minimum bounty value.": "Allocation value is smaller than the minimum bounty value.",
@@ -6,7 +7,7 @@
   "Bond is estimated based on bountyDepositBase and dataDepositPerByte constants.": "Bond is estimated based on bountyDepositBase and dataDepositPerByte constants.",
   "Bounties": "Bounties",
   "Calculated bond for bounty": "Calculated bond for bounty",
-  "Claim Bounty": "Claim Bounty",
+  "Claim": "Claim",
   "Claimable": "Claimable",
   "Curators deposit": "Curators deposit",
   "Curators fee": "Curators fee",
@@ -15,6 +16,7 @@
   "No open bounties": "No open bounties",
   "Proposer": "Proposer",
   "Proposer bond depends on bounty title length.": "Proposer bond depends on bounty title length.",
+  "Reject": "Reject",
   "Select the account you wish to propose the bounty from.": "Select the account you wish to propose the bounty from.",
   "The description of this bounty": "The description of this bounty",
   "The total payment amount of this bounty, curators fee included.": "The total payment amount of this bounty, curators fee included.",
