+++ packages/page-bounties/src/Bounty.tsx
@@ -11,7 +11,7 @@ import { AddressSmall, Icon, LinkExternal } from '@polkadot/react-components';
 import { BlockToTime, FormatBalance } from '@polkadot/react-query';
 import { formatNumber } from '@polkadot/util';
 
-import { BountiesActionButton } from './helpers/BountiesActionButton';
+import { BountyActions } from './BountyActions';
 import { getBountyStatus } from './helpers';
 import { useTranslation } from './translate';
 
@@ -58,7 +58,7 @@ function Bounty ({ bestNumber, bounty, className = '', description, index }: Pro
         <td>{beneficiary ? <AddressSmall value={beneficiary} /> : EMPTY_CELL}</td>
         <td><DueBlocks dueBlocks={blocksUntilPayout} /></td>
         <td>
-          <BountiesActionButton
+          <BountyActions
             bestNumber={bestNumber}
             index={index}
             status={status}
